Test on all supported versions of MRI Ruby and JRuby.
We wnumerate exact Ruby versions in .travis.yml. This is required for
at least Ruby 2.4 which isn't currently recognized as a fuzzy match in
Travis' RVM version. When specifying it as an exact version, it will
install it though.

We specify all the other versions too since this makes our tests more
reproducable over time.