Add hard gem dependencies to concurrent-ruby and ref
Generally, we try hard to avoid additional dependencies to external gems
to keep us from having to maintain all these dependencies in the variety
of environments where Rackstash is going to be used.

We still decided to depend on two concurrent-ruby gems since they are

1. are of exceptional code quality
2. are well-maintained and with devs eager to maintain a stable and
   well-understood interface
3. provide very useful buolding-blocks for safe interoperations across
   thread-boundaries.

The chosen versions are selected to be compatible with a wide range of
external frameworks.