Create new fields, tags and message objects on Buffer#clear
This helps in ensuring thread-safety when flushing unbuffered Buffers.
Previously, it was possibly for client code to still hold references to
fields on a flushed and cleared buffer, potentially resulting in
unintended side-effects if the client code was not aware that the buffer
was cleared.

By creating ompletely new object instances, existing references only
point to data from before the clear. This ensures a clean cut.