Simplify the tests for embedded IPs in the AnonymizeIPMask filter again
We still want to avoid the deprecated call to `IPAddr#ipv4_compat?`.
However, since we are always masking of at least one bit of any given IP
address anyways, we can ignore the specific tests for '::' and '::1' in
that method. The result of masking off these addresses will always be
'::' regardless of how many bits we are masking off anyways.