Enforce less restrictive version requirements for rake and bundler
With the release of Bundler 2.0, things got a bit messy. Since both
bundler and rake are very likely to retain backwards compatibility due
to their wide use in the wild, this is likely safe enough.

In any case, these are "only" development dependencies which do not
affect any production users of Rackstash but merely people running the
tests.