Update TruffleRuby in .travis.yml and allow failures again
Unfortunately, there are still various edge-cases where things randomly
break due to TruffleRuby differences or incompatibilities to other gems.

This time, it was an incompatibility with a json gem update described in
https://github.com/oracle/truffleruby/issues/1857.

We will still try to keep Rackstash compatible with TruffleRuby but will
not consider full support for it release-critical for now.